requirements.txt
aiohttp
annotated-types
ddapi>=0.4.7
nextcord==2.6.0
pillow==10.4.0
pydantic
plotly
PYYAML
orjson
numpy
